Implement `initializeAndRememberCredentials()`, enabling complex init scenarios
`WonderPush.initializeAndRememberCredentials()` not only initializes the
WonderPush SDK but also stores the given credentials so that next time
the SDK must initialize, it can do so on its own.

In complex customer scenarios, the credentials to be used cannot be
statically determined and are not available at startup, typically not
until the end-user selects a country or brand.

The WonderPush SDK however needs to be initialized in a few scenarios,
for instance when a push notification is received on Android, or
clicked. Because Flutter code does not always run in some of these
conditions, unless when using native code to handle the initialization
like recommended if those are constants, allowing the SDK to remember
the previous credentials allows it to initialize on its own.

Because the credentials are necessarily stored in a format readily
accessible for the application (although not to other applications), it
is necessary to opt into this behaviour by:

1. (Recommended) Following the native parts of the WonderPush Flutter
   SDK integration using `"USE_REMEMBERED"` for Client
   ID and Client Secret.
   - or - (Alternatively) Calling
   `WonderPush.initialize("USE_REMEMBERED", "USE_REMEMBERED")` in your
   `main()` method from `lib/main.dart`. You can alternatively use the
   value `null`.
2. Call `WonderPush.initializeAndRememberCredentials("…", "…")` with
   the real Client ID and Client Secret when the appropriate values
   are known.

Providing `null` values to
`WonderPush.initializeAndRememberCredentials()` will remove the
previously stored values.